created new function / react "like" component: NavbarComponent.
Because I want the standard page to be the login screen I have removed the navbar from index.html. So If I wanted to retrieve the navbar from navbar.html, I would have to load it first, im not sure I could load the html, while already on the home view, so I just created the NavbarComponent as a workaround.